Build tests with setuptools then remove it
Cython's internals rely on setuptools/distutils
and the wheel build uses an isolated environment so
the main project builds just fine.

Since we need to cythonize the tests though, we need
setuptools (which is not a dependency of cython itself).

Removing it after the build is to hopefully ensure this
project itself doesn't depend on setuptools.