Build test frames with ffmpeg
In order to test packing/unpacking for all of the FourCC's,
using a standard approach for test data fixtures seems
appropriate.

Generate packed raw data for each FourCC and include each
plane as a separate file using the `extractplanes` filter
(for YUV formats).

Include an archive of the data files to be extracted in
CI runners (so they don't need to have ffmpeg installed).